Add traverser pending count (see stopRunningJobs). Tweak Chain.IsDoneRunning again. Add Chain.FailedJobs (returns count of failed jobs). Update/fix tests. Fix proto.STATE_ (removing STATE_TIMEOUT shifted the byte values). Fix prepareSequenceRetry (real fix @todo). Remove SequenceTry and SequenceId from rm/joblog/store.

+++ request-manager/request/manager.go
@@ -166,7 +166,7 @@ func (m *manager) Create(newReq proto.CreateRequest) (proto.Request, error) {
 		jc.Jobs[jobId] = job
 	}
 	req.JobChain = jc
-	req.TotalJobs = len(jc.Jobs)
+	req.TotalJobs = uint(len(jc.Jobs))
 
 	// ----------------------------------------------------------------------
 	// Serial data for request_archives
